Fix flycheck running on a random downstream crate
Say you had A <- B <- C. If you save a file in A, the variable
`crate_root_paths` is populated with all three crates in that order.
However the one we actually ran flycheck on, was simply the first one of
A, B, or C to appear in `project.crates()`. Which is unordered.

This affected
- `[check] workspace = false` config where flycheck runs on individual
  crates
- non-Cargo build environments especially

Usually it was invisible because cargo prints out the warnings found
in A if you build B or C. But this is wasteful if B or C is a big crate.
So it was probably showing up as rust-analyzer constantly trying to
build random parts of your dependency graph on save.

Not every build tool does what cargo does either. Buck2's `[diag.json]`
subtargets do not contain warnings from upstream crates. That's how
I discovered this problem; warnings weren't showing up at all.